refactor(admin): supprime les commentaires inutiles et optimise les imports
- Nettoie le fichier `admin.ts` en retirant les commentaires superflus pour améliorer la lisibilité.
- Corrige le chemin des imports dans le fichier `recent-activity` pour respecter la convention de nommage (`activity-service.ts`).
- Prépare le module pour une meilleure maintenabilité et cohérence du code.
